+++ CHANGES.rst
@@ -47,6 +47,8 @@ Unreleased
 -   The return value from :meth:`cli.load_dotenv` is more consistent
     with the documentation. It will return ``False`` if python-dotenv is
     not installed, or if the given path isn't a file. :issue:`2937`
+-   Signaling support has a stub for the ``connect_via`` method when
+    the Blinker library is not installed. :pr:`3208`
 
 .. _#2935: https://github.com/pallets/flask/issues/2935
 .. _#2957: https://github.com/pallets/flask/issues/2957